requirements.txt
websocket-client>=1.6.0
requests>=2.31.0
PyYAML>=6.0
Pillow>=10.0.0
Flask>=3.0.0
pytest>=7.4